backend/requirements.txt
pandas
numpy
scikit-learn==1.6.1
xgboost
lightgbm
flask
flask_cors
imbalanced-learn
streamlit
python-dotenv
nest_asyncio
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
langchain-classic